refactor(api): separate profiles, portfolio and gallery modules
