+++ src/api/auth/routes.py
@@ -1,15 +1,10 @@
 import re
 from flask import Blueprint, request, jsonify
-from flask_jwt_extended import create_access_token
+from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity
 from api.models import db, User, ClientProfile, BusinessProfile, BusinessPortfolio, BusinessGallery
 from api.extensions import bcrypt
-import cloudinary.uploader
 
 auth = Blueprint("auth", __name__)
-
-def upload_image(file, folder):
-    result = cloudinary.uploader.upload(file, folder=folder)
-    return result["secure_url"]
 
 
 @auth.route("/register/client", methods=["POST"])
@@ -182,188 +177,17 @@ def login():
         "user": user.serialize()
     }), 200
 
-# ----------------------------------------------------------------------------------
-# BUSINESS PORTFOLIO
-# ----------------------------------------------------------------------------------
 
-@auth.route("/business/portfolio", methods=["POST"])
-def create_portfolio():
+@auth.route("/me", methods=["GET"])
+@jwt_required()
+def get_current_user():
+    user_id = get_jwt_identity()
 
-    business_profile_id = request.form.get("business_profile_id")
-    description = request.form.get("description")
-    logo = request.files.get("logo")
+    user = User.query.get(user_id)
 
-    if not business_profile_id:
-        return jsonify({"msg": "business_profile_id is required"}), 400
-
-    business = BusinessProfile.query.get(business_profile_id)
-
-    if not business:
-        return jsonify({"msg": "Business not found"}), 404
-
-    existing_portfolio = BusinessPortfolio.query.filter_by(
-        business_profile_id=business_profile_id
-    ).first()
-
-    if existing_portfolio:
-        return jsonify({"msg": "Portfolio already exists"}), 400
-
-    try:
-
-        logo_url = None
-
-        if logo:
-            logo_url = upload_image(logo, "business_portfolio")
-
-        new_portfolio = BusinessPortfolio(
-            business_profile_id=business_profile_id,
-            logo_url=logo_url,
-            description=description
-        )
-
-        db.session.add(new_portfolio)
-        db.session.commit()
-
-        return jsonify({
-            "msg": "Portfolio created successfully",
-            "portfolio": new_portfolio.serialize()
-        }), 201
-
-    except Exception as e:
-
-        db.session.rollback()
-
-        return jsonify({
-            "msg": "Internal server error",
-            "error": str(e)
-        }), 500
-
-
-@auth.route("/business/portfolio/<int:business_profile_id>", methods=["GET"])
-def get_portfolio(business_profile_id):
-
-    portfolio = BusinessPortfolio.query.filter_by(
-        business_profile_id=business_profile_id
-    ).first()
-
-    if not portfolio:
-        return jsonify({
-            "msg": "Portfolio not found"
-        }), 404
-
-    return jsonify(portfolio.serialize()), 200
-
-
-@auth.route("/business/portfolio/<int:business_profile_id>", methods=["PATCH"])
-def update_portfolio(business_profile_id):
-
-    portfolio = BusinessPortfolio.query.filter_by(
-        business_profile_id=business_profile_id
-    ).first()
-
-    if not portfolio:
-        return jsonify({
-            "msg": "Portfolio not found"
-        }), 404
-
-    data = request.form
-    logo = request.files.get("logo")
-
-    if logo:
-        portfolio.logo_url = upload_image(
-            logo,
-            "business_portfolio"
-        )
-
-    if "description" in data:
-        portfolio.description = data["description"]
-
-    db.session.commit()
+    if not user:
+        return jsonify({"msg": "User not found"}), 404
 
     return jsonify({
-        "msg": "Portfolio updated successfully",
-        "portfolio": portfolio.serialize()
+        "user": user.serialize()
     }), 200
-
-
-# ----------------------------------------------------------------------------------
-# BUSINESS GALLERY
-# ----------------------------------------------------------------------------------
-
-@auth.route("/business/gallery", methods=["POST"])
-def create_gallery_image():
-
-    business_profile_id = request.form.get("business_profile_id")
-    image = request.files.get("image")
-
-    if not business_profile_id or not image:
-        return jsonify({
-            "msg": "business_profile_id and image are required"
-        }), 400
-
-    business = BusinessProfile.query.get(business_profile_id)
-
-    if not business:
-        return jsonify({
-            "msg": "Business not found"
-        }), 404
-
-    images_count = BusinessGallery.query.filter_by(
-        business_profile_id=business_profile_id
-    ).count()
-
-    if images_count >= 6:
-        return jsonify({
-            "msg": "Maximum 6 images allowed"
-        }), 400
-
-    image_url = upload_image(
-        image,
-        "business_gallery"
-    )
-
-    new_image = BusinessGallery(
-        business_profile_id=business_profile_id,
-        image_url=image_url
-    )
-
-    db.session.add(new_image)
-    db.session.commit()
-
-    return jsonify({
-        "msg": "Image added successfully",
-        "image": new_image.serialize()
-    }), 201
-
-
-@auth.route("/business/gallery/<int:business_profile_id>", methods=["GET"])
-def get_gallery_images(business_profile_id):
-
-    images = BusinessGallery.query.filter_by(
-        business_profile_id=business_profile_id
-    ).all()
-
-    return jsonify({
-        "business_profile_id": business_profile_id,
-        "images": [image.serialize() for image in images]
-    }), 200
-
-
-@auth.route("/business/gallery/<int:image_id>", methods=["DELETE"])
-def delete_gallery_image(image_id):
-
-    image = BusinessGallery.query.get(image_id)
-
-    if not image:
-        return jsonify({
-            "msg": "Image not found"
-        }), 404
-
-    db.session.delete(image)
-    db.session.commit()
-
-    return jsonify({
-        "msg": "Image deleted successfully"
-    }), 200
-
-
